Naive Bayes for text data
its almost a sentiment analyzer - almost. Predicts whether the
'responses' in the dataset are either agreeing (positive) with a quote
or disagreeing (negative) with a quote.

Also includes entropy calculator for the scores. It gives out the
maximum and minimum entropy over all the records and their distributions
among the classes.